requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.0
Flask-Limiter==3.5.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.2.0
ordered-set==4.1.0
packaging==23.2
protobuf==4.21.12
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.0
typing_extensions==4.9.0
validate-email-address==1
Werkzeug==3.0.1
wrapt==1.16.0
